fix(release): unbreak CI for 0.3.1 — mock, TOML, clippy ICE, fmt
CI had been failing on every push since v0.2.0 (only the v0.3.0 tag
release pipeline succeeded because it skipped the workspace gate
that was actually red). Three distinct regressions had piled up:

1. Headless mock provider was no longer parsing the peridot action
   envelope into native tool_calls. After the native-tool-call
   refactor (4b2d4c8), FileMockProvider just returned the JSON as
   raw text. Result: agent_done summaries appeared in
   `tool_result.summary` as the literal action JSON, and the
   harness never executed the mock-described tool, so budget /
   max-turns fixtures all stopped at Done with exit 0. Fix:
   parse_mock_completion_response now emits ToolInvocation for
   both top-level {action, parameters} and text-envelope-wrapped
   variants (the latter carries usage metadata for budget tests).

2. default_offload_threshold_chars returned usize::MAX which is
   u64::MAX on 64-bit. `peridot setup` writes that to TOML, but
   the TOML spec defines integers as signed 64-bit so it could
   not be parsed back. Switched the sentinel to i64::MAX as usize,
   still effectively unbounded (~9.2 × 10¹⁸ chars) but round-trips
   cleanly.

3. clippy 0.1.95 ICEs inside `needless_borrows_for_generic_args`
   while rendering a diagnostic against peridot-tui ("slice index
   starts at 45 but ends at 44" in annotate_snippets). Added a
   crate-level `#[allow]` with a comment pointing at the upstream
   bug; harmless once clippy ships a fix.

Also applies `cargo fmt --all` (the format gate had drifted across
prior commits) and replaces a manual `Default for BranchPickerState`
impl with `#[derive(Default)]`.

QA: cargo fmt --check ok, cargo clippy --workspace -- -D warnings
ok, cargo test --workspace 339 passed / 0 failed, cargo test -p
peridot-cli --test e2e --features e2e 11 passed / 0 failed.

+++ peridot-common/src/lib.rs
@@ -911,7 +911,11 @@ fn default_offload_threshold_chars() -> usize {
     // Effectively disabled. See peridot_context::ContextLimits::default() for the
     // rationale: modern models support large enough contexts that disk offload causes
     // more harm (recursive re-reads on smaller models) than memory benefit.
-    usize::MAX
+    //
+    // `i64::MAX` (≈9.2 × 10¹⁸) is chosen instead of `usize::MAX` so the value
+    // round-trips through TOML, whose integer spec is signed 64-bit. Still
+    // larger than any realistic observation size.
+    i64::MAX as usize
 }
 
 fn default_observation_max_chars() -> usize {
